Change how importer_program.py is executed.
If app is not frozen, we use the python interpreter that was used in starting the app. Previously we used the
Python interpreter that was set by user in Settings->Tools.

If app is frozen, we need to prepare python.exe that is EMBEDDED to the application
installation directory (/extras/python.exe) for running this script. We need to manipulate it's
PythonPath before importing spinedb_api in order to make sure that we import the
spinedb_api that is included in the installation bundle.

The bug that this commit hopefully fixes was discussed in Issue #603.