Get parsed value from model also when plotting
In previous commits we introduced a minor optimization which consists
in holding the parsed parameter value (i.e., a python object of type
Map, TimeSeries, etc) in the Qt.UserRole. With this, parsing is
effectively done as little as possible throughout the app.

In this commit we realize plotting can also make use of the above.