Fix bug in updating db values
We were updating items in cache *after* sending the corresponding
signal to the listener DS forms, which caused incorrect data
display.

The reason was we wanted to have a last look at the items before
definitely updating the cache, to do some work on the filter menus of
parameter tables.

But now we keep a dedicated 'cache' in the CompoundParameterModel only
for this purpose.