Don't plot all text x tick labels
In tabular view when plotting expanded indexes it is possible to have a
large number of text indexes. As with Map plotting, we need to limit the
number of indexes drawn on the plot window because otherwise matplotlib
draws every single x tick label making the plotting very slow.

Re #520