Fix bug in setting data with the ParameterValueEditor
ParameterValueEditor was relying on calling setData with the given
model index to update values in the database. But model indexes are
not persistent in our system; the user can remove or add rows, or
reset entire models as in the case of the pivot table, *all while
keeping the ParameterValueEditor open*.

So now we implement an alternative where the ParameterValueEditor
talks directly with SpineDBManager, which seems to work fine.