Introduce Load template and Save as template actions in DS form
Re data#86

The idea as of now is to use JSON templates to initialize databases.
For example, Spine Model will have it's own template that the user
can load into an empty database to initialize it, or even into a filled
database to update parameter values and default values (nothing more
complex than that).

The check box "for Spine model" is removed from the DS properties
consequently.