Use the clean status of the undo stack to check if dbs are dirty
....instead of db_map.has_pending_changes(), which behaves a bit weird
with undo/redo. For example, we may add an item and then undo,
and `has_pending_changes()` will return True in circumstances the db
is essentially clean.

Maybe the ultimate solution is to move the undo stack to spinedb_api
as we discussed over beers and whiskey in Espoo 2020. In this way
we could also get rid of the diff tables; the only 'problem' is that
for rolling back we'd need to undo the entire stack which feels a bit
'risky'.