Move tool specifications to main toolbar and make them draggable
Re #650

This is just to see how it would look to have lightweight plugins
int some near future. So now, tool specs are some sort of plugin
build on top of the Tool item, that behave similarly to project items
in the sense they can be dragged and added to the project directly.