Fix issues with Data Connection renaming
The Data files list was not updated correctly immediately after renaming
a Data connection item. We now call refresh in the rename() method to do
it explicitly. Also, there was a superfluous call to refresh() in
restore_connections() resulting in unnecessary file system accesses.
This has been removed as well.

Re #620