Fix rename_dir() when target already exists
shutil.move() used in rename_dir() does not raise FileExistsErroro if
the target is an existing directory, contrary to what the code expected.
Instead, shutil.move() moves the destination inside the target
directory. This is not a renaming operation as the function's name
suggests. We now check the existence of target directory before calling
shutil.move() and bail out if it exists.

Re #620