Fix ok_to_close() in SpineDBManager
Checking if the undo stack is not enough to know if there are pending
changes in the database map. This was causing a situation where closing
the Data store view was not possible because database manager thought
that there were pending changes while the database map failed to roll
back the session because those pending changes did not exist.