Rework item selection in Data store view
DataStoreFormBase was manually making sure that only one of the
tables/lists/trees had selections. The idea behind this was that this
way the 'target' for any action (copy, paste,...) was clear and could be
read from a single attribut. However, this posed problem e.g. in the
Graph view where the graph was cleared when anything else than objects
or relationships were selected from the tree views.

This solves the issues by dropping the entire selection tracking mechanism.
Multiple views can have selections at the same time. The 'target' widget
is chosen by which widget has the focus. A focused widget ought to be
the one user interacted last and most probably wants to be the target of
the action as well.

Re #627