Don't export impossible parameters, sets
Gdx exporter was exporter all parameters. This does not make sense if
the parameters dimensions, i.e. the indexing domains were not exported
as well. We now drop parameters that will not have an indexing domain in
the .gdx file.

Further, Gdx Settings window now explicitly disables exporting of sets
that are indexed by domains that will not be exported.

Re #668