Fix Exporter item_dict
A previous commit introduced a new attribute for gdx indexing settings,
set_name. This broke the backwards compatibiliy of Exporter's item dict.
In fact, we don't need to save set_name to JSON as it can be read at the
same time as we check that the parameter is still valid for indexing.

Re #674