Fix bug in duplicate object
Re #271

We shouldn't have relied on "formatted_value" to get things going,
but we thought we could get away with the it. The problem is one needs
to make the value show in the parameter panes or pivot table for the
"formatted_value" key to be created.

Of course we could always read the right value using the
SpineDBManager method (I think is `get_value` with role=PARSED_ROLE.

But now we're trying what seems like a more consistent approach
based on SpineDBParcel. Hopefully the bug is gone.