Improve Importer properties tab usability
The 'Import Editor...' button was very grumpy if nothing was selected in
the Source files list. This was annoying if there was only a single
item. If nothing is selected, we now select the first eligible item on
the list when 'Import Editor...' is pressed to go with that.

Re #597