Fix .csv connector preselection
There is a dialog asking which connector should be used for import
mapping before the Import Editor is opened for a previously unknown
file. File extension is used to preselect an appropriate line from the
list of choices. However, .csv, .dat and .txt files failed to preselect
the CSV connector because the if clause was slightly wrong dropping a
row index of 0 as False.

Re #597