Add option to attach exported SQLite file to a Data Store in the project
Re #570, #518

I just want to check if something like this is useful.
After exporting an SQLite file from the DS form, the notification
bubble has a link that the user can activate to attach the file
to a Data Store in the project.

This can serve to create Data Stores with incremental changes
that then can be merged in the Recombinator.