Centralize scene resizing in one slot
We now provide a slot called `_set_preferred_scene_rect` in our
`CustomQGraphicsView` class, that we can call to set the scene rect
to a preferred one. At the moment it is called when
shrinking-if-needed, after zooming, right after loading a new
project, and after ensuring item's visibility.

The preferred rect is computed by uniting the viewport and the
items' rect. It works best in my opinion, since it guarantees
the scene won't shift after dropping an item, for instance.