Fix zoom and resizing issues in Design view plus some renaming
We explicitly use the item's bounding rect to determine the fitting
zoom, since the scene rect is now kept just as big as the viewport.

We also set the preferred scene rect after resizing, but we use the
time line to do it only every 200 msecs. This led to some renaming
in order to clarify what's what.