DataStoreForm: refactor context menu of entity trees
The context menu handler was relying on a block of if-else
to react to the selected option. Didn't look so well on code, so
now I'm using a different approach which consists on connecting
the triggered signal of the menu actions.

In addition, we reuse the same menu instead of creating a new one
every time.