Fix importer restoring selected mappings
Mapping selection state was stored in project.json as a list of boolean
values. This approach was broken, however, because there is no
one-to-one correspondence between this list and the settings list. Some
labels may be missign in the settings list because user never defined
any mappings for these.

We now store the selection states as a pseudo-dictionary where each
boolean state has a corresponding label so the selections can be
restored independently of whether the labels are in the settings list or
not.

Re #700