Refactor parameter table filtering
Hopefully it's more readable now. The previous implementation grew out
of several patches as we introduced more functionality and tried to
solve bugs.

But now all the filtering is done in the ParameterViewMixin, so it's
all in one place. Also, we no longer keep separate filters for
different entity types as it was too cumbersome and not too
advantageous from the performance point of view.