Store the graph view items in lists so they don't disappear
For some reason, it *seems* that Python was garbage collecting some
of the items in the graph view causing them to vanish.
To prevent this, we just store all the items in a list in
the GraphViewMixin. This is actually not dramatic and gets around
the issue, so it's ok.