Fix problem with undoing items removal in graph view
When removing items and the undoing, the graph was bringing back
the items together with all their relationships, some of which weren't
originally in the graph, leading to confusion.

The above was an undesired side effect of materializing relationships
added from the graph. Now we explicitly differentiate between the
two situations.