Refactor pivot table model with a more pythonic approach
The previous version of the model was constantly checking the
'input type' of the parent Data store form to know how to show the
data, how to edit the data, and other stuff.

Here, we got rid of those if elif else blocks in favor of a subclass
approach. Basically we have one PivotTableModelBase subclass for each
input type which greatly clarifies things. The model is constructed
and installed on the view every time the user changes input size,
which seems to work nicely.