Fix error in defining the julia project path
I was under the impression that using --project=@. activated the
home project, but that's not the case, it actually activates
the local project at the spine toolbox's app folder.

So I'm trying to fix that now, at least works better for the Julia
console, haven't tested it for the shell.