Fix updating tool specs
Re #718

Whenever a tool spec is edited, we need to update in all the tool items
that use it. For this we were at some point holding the category in
the specification just so we can retrieve items more effectively from
the project item model. But when we moved from category to factory we
thought we could also get rid of the category attribute in the
specification with no consequences. At that point we broke it.

Now it should be better.