Manually set the cursors we want in mouse release event
Re #719

It's difficult to call the base class implementation and at the
same time get the right cursors. But on the other hand, manually
achieving what the base class implementation does is nonsense.

So we now use a carefully crafted condition in the mouse release
event of the design view to adapt the cursor at best.