Fall back to matplotlib.tri in case scipy.interpolate not good
Re #717

This is rather a desperate attempt to get around `ImportError` on
`import scipy.interpolate`. Apparently, it can be sometimes difficult
to import that module under certain Windows configurations.

I'm unable to reproduce the bug at the moment (haven't tried, really),
so I can't guarantee this is of any help. But let's try.