Filter filter menus when filtering the underlying parameter table
Re #679

This is so the filter menus only present values that are actually
shown in the table. Since the table can be filtered, the filter
menus should also be filtered, like inception, and this is what we
do in this commit.

The approach is mainly brute-force, we don't try to optimize by
keeping maps of values to rows or whatever. We just sweep the lists.