Trying to open a project with unknown item types does not crash Toolbox anymore
In case there was an uknown item type in project.json, a tuple was returned instead of a
dictionary which was requested by the callee.

It now loads all project items that are recongnized and prints an error
message with the project item name that could not be loaded. Previously
it loaded all project items until an unrecognized item type was encountered
and stopped there, which was confusing since there could be more recognized
project items that could be loaded.

Restoring a connection failed message now says the source
and destination project item names.

Resolves Issue #697