Catch exceptions when names clash
GAMS is case insensitive which makes it possible to have duplicate set,
parameter or scalar names when writing a .gdx file. In these
cases, gdx2py raises an exception which was not caught by the gdx module
causing Exporter to fail silently. This adds try-except blocks to where
they are needed.

Re #738