Cleanup importer mappings
Import editor was trying its earnest to do 'the right thing' with
mapping references. Unfortunately, the code was riddled with isinstance
checks that made the it very fragile. The functionality has been now
moved to spinedb_api side where it actually belongs making it possible
to simplify the reference manipulations on Toolbox side.

Re #733