Fix lambda again
Re #743

Here is how I think it works: we are connecting the action's
`triggered` signal to the given slot. So the first positional argument
should be the `checked` argument from the signal, and then we
introduce our `item_type` keyword argument.

No need for the `specification` keyword argument as we are not
using it in the call to `parent.show_specification_form`.