Fix crashing DB fetch thread
Fetching a database while opening a Data store view sometimes abruptly
killed Toolbox. It seems that this was caused by destroying the fetcher
by the calls to self.done.emit() and self.deleteLater() in
SpineDBFetcher.start(). Because start() was called by the GUI thread,
the fetcher object could be destroyed before the worker thread had
finished processing the slots.

We now explicitly queue a signal in the worker thread that triggers the
fetcher's clean up and removal in the GUI thread.

Re #749