Fix db fetcher thread calling UI stuff
Qt warnings were being printed in the console when alternatives and
scenarios were fetched during opening the Data Store view. This was due
to the fethced thread trying to show notification windows to the user.
This, in turn, was because the fetcher was calling the receive_*_added
methods of TreeViewMixin instead of the correct receive_*_fetched. The
porper methods are now implemented in TreeViewMixin and the warnings
are gone.