Support arbitrary JSON value in parameter value lists
Re #755

It is now possible to define parameter value lists using any of our
JSON types, which is nice.

Right-click on any value list item in the parameter value list tree
to bring the context menu, and then select Open in editor to bring
the Parameter value editor for that item.

Editing directly by typing also works, as before.