Fix parameter value delegate
Re #755

When editing parameter values that have associated a value list,
we need to present options using the SearchBarEditor within the
delegate. This kinda forces us to do the parsing of value list items
in SpineDBManager so it's available from everywhere (rather than doing
that parsing in the ParameterValueListModel, as before).