Start refurbishing spine datapackage editor
Re #705

Basically, I'm re-purposing this widget as a decent generic CSV editor
with support for creating a datapackage on the go.

The functionality to do import/export between a datapackage and the
Spine format has been removed for now. In case we need it back, we just
need to track this commit. Perhaps we can find a way to implement
such functionality within the Importer project item. But we need to
find a place for it and since it's not in use, removing it
(temporarily or permanently) seems wise in order to reduce noise.