Fix context menus in Data store view
The context menus of the tree view in Data store view were broken due to
the alternatives and scenarios tree view not containing action related
to relationships. Also, the context menu of the object tree view did not
show up because of an issue in the .ui file. EntityTreeView base class
has been now split into a lower-level ItemTreeView to accommodate the
scenario and alternative tree view into the tree view hierarchy better