Refactor Import preview window UI with QDockWidgets
This is so the Import preview doesn't look and feel too different
from our other main windows. It also requires fewer UI files as
ImportEditorWindow.ui already embeds mostly everything it needs.
The mixin strategy is used to keep controller files separated
as it was before, which is nice for isolating functionality.