Fix adding parameter values in DS form
Parameter values weren't being added, just displayed. With this
change we correctly find the alternative_id from the alternative_name
entered by the user, to send the parameter_value to the db.

We also make the field `alternative_name` fixed because it doesn't
make sense to modify it after the parameter value is in.