Don't combine alternatives and scenarios across dbs
If we do it, it becomes impossible to associate one alternative
to a scenario in *only* one db. It doesn't work nicely with the drag
and drop approach. Something similar happens with the "active"
checkbox.

So now we basically split alternatives and scenarios across databases,
similar to what we do for parameter value lists.