Fix bug when undoing db remove operations
For example, if we'd remove a tag and then undo the operation, the tag
wouldn't be reattached to the parameter definitions.
Same for scenario alternatives.

With this commit, we start managing a cache of mapping-tables
(namely, parameter_definition_tag and scenario_alternative) just for
the purpose of being able to readd removed items.

The whole `SpineDBManager.remove_items` should be optimized too,
as we were performing too many dict lookups, from id to data, then back
from data to id, because it was a bit disorganized. Not that those
lookups are really expensive but who knows.