Revert undo for mapping selection changes
In the Mappings list in Import editor, it was possible to undo/redo even
item selections on all lists. This was to get 'perfect'
undo/redo so we could rely on every action being recorded in the undo
stack. This quickly became too tedious, however. This commit drops the
SelectMapping command simplifying item selection in the Mappings list.
We now try to change the selection automatically in a smart way during
undo/redo to keep the user informed on what is changing.

Re #454