Make item ranks show relative execution order
Since the move to dagster, we have been no longer able to rank project
items truthfully. This tries to circumvent the issue by changing the
ranks from absolute execution order to relative one: the ranks show the
correct order when following a continuous 'path' in the DAG; however,
items in different parallel branches may have the same numbers so it is
impossible to tell in which order the parallel branches are executed
relative to each other.

Re #548