Prevent gdx from trying to expand scalars
There were cases when the gdx module tried to apply indexing settings on
a scalar parameter. This was possible e.g. when the indexed parameter
was in an object class while a scalar parameter with the same name lived
in a relationship class. We now check that the parameter has the correct
original dimensions before proceeding with index expansion.