No need to handle ValueErrors anymore when creating executable items
Now the Gimlet's ExecutableItem constructors doesn't raise,
but proceeds so the execution fails when reaching it. This is more
consistent with how the other project items work. In the future we
may want to work around different situations in the `execute` methods,
but at the moment it's ok to go with this imo.