Upgrade spine_items from non-installed versions of toolbox
Re #845

The idea here is to first download spine_items into a temporary
location, look at the REQUIRED_SPINETOOLBOX_VERSION
and then decide whether or not to install it.

With this, users of toolbox dev branch don't need to run
pip install to make the automatic upgrade of spine_items work.