Add support to create a 'silent' execution_item
This is because we use execution_item() in
resources_for_direct_successors() so we can call
_output_resources_forward() instead of duplicating the same code.

But the undesired side effect is that we get a lot of execution
related event messages that we don't want. We want to know what
resources are being passed only.

So this commit remediates that situation by making the result of
`execution_item()` silent by default.