Fix arrow filters to use scenario names rather than ids
Using the scenario id for a scenario filter doesn't work well
with the writing filter. The writing filter is applied to the
downstream resource because we want to write to a scenario from
one of the upstream resources. This means that generally the
scenario doesn't exist in the downstream, so the filter needs
to create it. Anyways, this is impossible if we use the id instead of
the name. So now we use the name without too much complications.