Fix traceback when showing the context menu with remove items
Re #884

When we remove items from the db, we don't remove them from the
entity graph scene. instead, we just hide them and put them in
a list. The point is that if the user undoes the removal operation,
we can re-show the items back in the exact same position.

Now, sometimes we need to loop over all items in the scene to
collect information (like when we're about to show the context menu).
So we need to skip the items that are no longer in the db, and this
is what we're doing with this commit.