Use temporary directory for project in tests
We don't create a fixed project directory anymore for unit testing.
Instead, tempfile.TemporaryDirectory() is used. Unit tests don't need to
mock create_directory() anymore when building a ToolboxUI and a project
allowing more realistic testing of projects.

This brought up an issue in CustomFileSystemWatcher which was not
cleaning up properly in tear_down().