Fix undoing critical commands (item renaming) on project closing
Re #664

We were missing some stuff, but the idea was already there.
When closing the project, we undo all critical commands.
We can specify for each command whether or not is critical using
``SpineToolboxCommand.is_critical()`. At the moment, only
item renaming is critical.

If undoing any critical commands is not possible (e.g.,
directory name clash because somebody is messing around with
the project folder), we don't close the project so the user can
fix the situation.