Fix pivot table when entities are added/removed
IndexExpansionPivotTableModel failed when an object or a relationship
was added to or removed from the database. This was caused by the index
expansion model using additional fields to identify data on the pivot table
compared to what ParameterValuePivotTableModel (the parent class) was
trying to use. IndexExpansionPivoTableModel no more delegates handling of
database modifications in these cases to its parent class.

Re #970