Use subprocess.run() to restart toolbox after upgrading spine_items
Re #968

This seems to work a bit better on windows, since the control doesn't
return to the prompt as with `os.execl()`. I'm not aware of any
big disadvantages. I wonder if it's more expensive in terms
of resources. It doesn't seem like a big deal though,
as most of the times toolbox starts without upgrading items...