Fix bug when adding relationship parameter values
Weird implementation error: we were consuming the same set of rows
twice, for adding parameter values and then relationships on the fly.
But the second time the situation in the table had changed
(some items were already moved into the 'full' table),
so some of the rows were no longer valid --> IndexError.