Smart Edit action enabling & disabling
- Edit actions (copy, paste, remove, etc.) are now enabled/disabled depending on user actions and
project state. For example. Paste is only available if there is a Project Item on the clipboard, Remove
All is only available if there are project items in the project, etc. This affects the main window File->Edit
menu, project tree view context menus and context menus on Design View. Also enabled Tooltips for
these menus.

Resolves Issue #1213