Use julia default project in kernel editor
The julia default value for the `--project=` switch is the empty
string, which means use the project at .julia/environments/<version>

In the kernel editor we were using `@.` as default, which means use
the project at the current working directory. The result in this
case depends on where toolbox is started from, which is difficult
to manage. So I just changed to follow the julia natural behavior.