Disable project related actions if no project open
This is so the user can't add any items or specs before opening a
project. Before, we were handling this by printing a message to the
event log, 'Please open or create a project first'. I believe
this is more robust and consistent with more recent developments.
If an action is impossible, we disable it so the user is not
surprised when they activate it and it does nothing.