Don't build SpineDBEditor ui components in another thread
It was causing the app to crash once in a while.
At some point this warning was printed

QBasicTimer::stop: Failed. Possibly trying to stop from a different thread

We need to investigate more. In the meantime, better build the UI
in the main thread where widgets are supposed to be.

Querying the database is still done in the other thread, which is
good.