Use as few QSvgRenderers as possible in the Entity graph
This is not related to any issue, but it has to do with
making the DB editor's entity graph completely svg-based
(rather than pixmap) based. This improves performance (I think),
visuals, and most importantly it allows 'Export as PDF' to generate
a vector pdf (not raster) which is optimal for embedding into latex
and stuff.

I think this should be one of last commits on this. At the moment
we have a nice setup but you never know.