Queue Spine db fetchers so they never crash
Since all fetchers need to run in the same thread (the one that
has the db map objects) we can queue them. This allows us to
manage signal-slot connections more carefully, and avoid any risk
of segfaults.

However, it is very unlikely that a large number of Spine db editors
are fetching at the same time, but you never know.