Fix traceback in Data store's undo notifications
There are widgets in Data store that get destroyed before the undo
notification widget associated with them get closed, e.g. the line edit
that is used to add scenarios. When undo notification widget closes it
tries to call the focus widget but fails since the widget has been
deleted on C++ side.

Undo notification widget now listend to focus widget's 'destroyed' signal
to know if the widget is already gone and should not be used anymore.

Re #1285