Refactor item updates, Design view
Changes in project items used to trigger Toolbox to crawl through the
entire DAG in two directions. This is rather time consuming process and
attempts to optimize it caused issues such as #1276. This is a
refactoring of that machinery.

We now try to update only affected items when a project item is updated
or the DAG changes. This hopefully leads to more performant updates. To
achieve more fine-grained updating, refactoring of how the project is
edited on the Design view needed to change. Project's undo stack
commands now talk directly to project when items or connections are added,
deleted etc. Project, in turn, signals the Design view to update the GUI
accordingly.

Re #1276