Fix security vulnerability in a dependency
A vulnerability was discovered in the PyYAML library in versions before 5.4, where it is
susceptible to arbitrary code execution when it processes untrusted YAML files through
the full_load method or with the FullLoader loader. Applications that use the library to
process untrusted input may be vulnerable to this flaw. This flaw allows an attacker to
execute arbitrary code on the system by abusing the python/object/new constructor.
This flaw is due to an incomplete fix for CVE-2020-1747.