Make Map editor less eager with DateTimes
DateTime constructor is very eager to convert any string into a time
stamp. This causes issues when pasting data into Map editor because many
strings that are not meant to be date times get converted into such. We
now use some heuristics to identify strings that we know are not time
stamps to prevent too eager conversion.

Re #1108