Speed up item removal in Spine DB editor
Basically, removing all items and then undoing was adding a lot
of items that weren't needed for the UI. This is because removing
needs to consider the entire db, not only the visual part.

So now, instead of creating the UI for the re-added items,
we put them into the fetching cache so they are fetched as needed.