Sort items in entity tress upon insertion
Re #1455

Because of lazy loading, we may be inserting items into the entity trees
in a different order than they come from the query. In detail, this is
just because sometimes when we progress the query iterator, we find
results that don't need to be shown right away. In this case we store
these results in a fetching cache, so we can create the corresponding
UI elements later on, when it's needed. And this can unsort the items
in ways difficult to predict. So now we use `bisect` to keep entity
tree items always sorted.