Improve can_fetch_more implementation in DB fetcher
Re #1454

Before, can_fetch_more was unaware of the fetch success condition,
so it was returning True in cases where there weren't any items to
fetch. For this reason, empty relationship classes weren't grayed out
in the entity tree view.

Now we use the success condition, but checking the condition is
expensive and the function is called very often, so we also memoize
results.