Improve entity group fetching
Before, entity groups were fetched by the entity class item, but this
didn't work well if the group had many members. We need to fetch
groups in the parent item for everything to work well.

Also, SpineDBFetcher.fetch_more() was failing because of a
de-synchronization between the UI and DB threads. Sometimes,
the UI thread would look into fetching cache and the DB thread
would write into the fetching cache at the same time. We handle this
by doing the entire fetch more thing in the DB thread (also the part
where we look into the cache).