Fix management of actions ProjectItemSpecArray
Re #1470

We need to the `_actions` attribute to be a dict mapping rows
from the specification model to the corresponding QAction in the array.
A list doesn't work nicely because the row numbers come from different
slices in the spec model (they don't always start at zero).

This fixes a traceback when editing specs.