Make sure undo stacks exist in DB worker
When closing the Database editor with uncommitted changes, it was
possible to run into a race condition where the worker thread tried to
clean DB manager's undo stacks which had already been deleted by the GUI
thread in SpineDBManager.close_session(). We now make local copies of
the undo stacks in SpineDBWorker to make sure they remain accessible in
the worker thread.

Re #1486