Fetch db contents broadly for filling up editing widgets
This makes sure that all the data is 'offered' when users want to
add/edit/remove content. So as it currently is, only 'viewing' data
is lazy, everything else is not lazy as it's a nightmare to make it
lazy, and not really needed since we need to fetch the entire db
to make integrity checks anyways.