Move user prompts from DB manager to editor
When removing listeners, Database editor was prompting user if they
wanted to commit/roll back dirty database. A message box was used for
this and it was parented to the currently active window which presumably
was the DB editor. However, this was not always the case making DB
manager to produce a Traceback. This commit moves user prompts for
committing/rolling back from DB manager to DB editor making sure we
parent the dialogs properly.

Re #1496