Add the value_index_list to paramter_value_list items
Re #1508

The problem here was that we were not correctly replicating the
'query' result when adding parameter value lists. Our strategy is,
instead of querying the db again to get the 'extended' version
of the item (with all the fields), we replicate that version
'manually' by getting stuff directly from the cache (faster).