Move execution list closer to console and log
THe execution list was begin shown at the bottom of the item
properties which made it difficult to find, and also possible for it
to be not visible. Now we move it closet to the console and item log
so the user cannot possibly miss it. This means though that we need
to use two widgets, but it's worth it.