Fix CustomFileSystemWatcher
CustomFileSystemWatcher._handle_dir_changed() took a snapshot of the
changed directory and compared it with a prevous snapshot. The method
was assuming that only a single file was added, removed or renamed but
in reality there could be more files involved. We now properly check all
changes in the directory's files.

Re #1587