Prevent opening new tabs while window is closing
It was possible to open a new DB editor tab in a multitab window that
was closing. This resulted in a situation where the new tab never
appeared but was still registered as a db manager listener. Reopening
the database in a new editor window would trigger notifications to the
closed tab as well leading to Tracebacks as the old window's widgets
were already deleted in C++ side.

MultiTabWindow now has a flag that prevents new tabs to be created or
inserted. Also, DB manager checks for this flag before deciding if a new
editor tab should be created in an existing or new window.

Re #1570