Don't send connections or jumps outside the dag to engine for exec
There was a problem when running a dag while having a jump in
another dag. Engine was complaining that it couldn't find the source
or destination item for the foraneous dag.

To solve this, we restrict the connections and jumps we send to
engine to the ones in the actual dag being executed.