SpineDBEditor: Fix inserting models with no rows in parameter model
Re Spine-project/Spine-Toolbox#1693

Sometimes if the user has a selection in the entity treeview and then
adds data into parameter tables using the empty row, the newly added
data is not visible because the treeview selection filters it out.
In this cases, the model 'sees' an insertion with no data. In general
when this happens, we want to trigger fetching because there might
be more data in the database that needs to be shown. We do this
by emitting layoutChanged.

At the same time, when the user adds new parameters the 'empty' rows
(where the user entered the data) need to be removed. Calling
removeRows in the same loop cycle as emitting layoutChanged causes funny
things to happen. Typically more rows than needed are removed from
the 'empty row model', but I also saw some crashes.

The solution is to delay either layoutChanged or removeRows to the
next cycle by using a QTimer.