Persistent console: make the text buffer flush timer single shot
I'm not sure, but I believe the Slot that flushes the buffer was
being called while executed, leading to segfaults. It would be good to
confirm that calling a Slot while being executed causes segfaults.

So, my suspicion is sometimes, it took long enough to flush the buffer
for the next interval to elapse, and the Slot was called again. Now
we make the time single shot, which means there's no chance for it
to timeout while the previous timeout was being processed.