Fix rare Traceback in alternative/scenario tree
It was possible to get a Traceback when dragging and dropping
alternatives in alternative/scenario tree. The issue was difficult
to reproduce but seemed to involve scrolling the tree in too short
dock area and then dropping the alternative below the scenario tree.
Usually, a drop like this should be refused but for some reason
(bug in Qt or canDropMimeData?) it was sometimes accepted causing
a Traceback in AlternativeScenarioModel.dropMimeData() as
scen_alt_root_item's type wasn't what was expected.

Re #1109