Fix line number highlighting with wrapped text
CodeTextEdit didn't properly update the line number area when the cursor
was on/was moved from a long line of text that had been wrapped. This
happened because _update_line_number_area() couldn't pass a large enough
update rect to LineNumberArea.update(). We now specifically listen to
changes in cursor position so we can feed update() with the correct
bounding rect. The change also allows us to ignore regular but completly
unnecessary calls to update() due to cursor blinking.

Re #1750