Fix parameter value table delegates
When parameter table delegate needed database mapping in Database editor, it would use the database codename in the last column. This would fail if there were multiple databases with the same codename. The delegates now get the correct database mapping directly from the model instead of using the codename fixing the issue.

Re #1764