Don't connect DB editor's copy and paste actions
Database editor's widgets use the editor's copy and paste actions. DB
editor also handles the trigger signals of the actions. Despite that,
CopyPasteTableView was still connecting the signals to its own copy and
paste slots causing the actions to trigger twice. We don't connect the
actions' signals anymore to prevent this.

Re #1762