Fix adding data to pivot model
It wasn't possible to replace missing values, i.e. Nones, in PivotModel
because the data-to-be-added was updated by the existing pivot data.
This made e.g. pasting data to the pivot table impossible. We now update
the existing data by the added data properly while making sure that it
is still possible to add missing data, too.

Re #1762