Refresh copy/paste action upon selection
Problem: Database editor's copy and paste actions were disabled until a
suitable context menu was shown preventing the usage of the usual
keyboard shortcuts before a menu was shown.

Solution: we now refresh the enabled state of copy and paste actions
when anything is selected in tables or trees making the actions
independent of context menus.

Re #1762