Fix row/column appends in MapModel
MapModel now inserts a new row/column to the end of the current
rows/columns when the table is extended by the expansion region.
We used to insert the row/column between the actual data and the
expanse region which resulted in the current (selected) cell jumping
around unintuitively in Map editor.

Re #1802