Fix dirty database committing when editor is closed
Database editor used to check if mapping was orphaned and if so, prompt
if changes should be committed before closing the editor. This worked fine
as long as all listeners were Database editors. However, nowadays we may have
listeners that are never "closed" and cannot actually make commits such as
filtering connections.

This implements a more fine-grained check if changes should be committed when
closing a database editor. Database listeners can now declare themselves
as editor as well, and instead of checking if db mapping has remaining listeners,
we check if it has remaining editors.

Re #1810